requirements.txt
aniso8601==3.0.2
Django==2.1.1
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-graphql-jwt==0.1.12
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
promise==2.1
PyJWT==1.6.4
pytz==2018.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
sqlparse==0.2.4
typing==3.6.6
